+++ modules/pico_http_util.h
@@ -108,7 +108,7 @@ struct pico_http_uri
 
 /* used for chunks */
 int pico_itoaHex(uint16_t port, char *ptr);
-uint16_t pico_itoa(uint16_t port, char *ptr);
+uint32_t pico_itoa(uint32_t port, char *ptr);
 int8_t pico_processURI(const char *uri, struct pico_http_uri *urikey);
 
 int is_digit(char chr);
